Flask/requirements.txt
flask==2.2.3
flask_sqlalchemy==3.0.3
pillow==9.5.0
mysql-connector==2.2.9